modify code style
